xfs_repair: preserve in-core meta_uuid while zeroing unused sb portions
If we zero unused parts of the superblock, we must preserve meta_uuid
across the zeroing because meta_uuid is used to verify the v5 format
checksums even on non-metauuid filesystems.  If we don't, the next
thing that happens is that all metadata fails in the verifier and the
whole filesystem is "toast".

Signed-off-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Dave Chinner <david@fromorbit.com>