xfs_quota: fix free command for foreign fs
The "free" command is really just a fancy df that knows about
log space and realtime blocks for an xfs filesystem.

We can simply use statfs to get more or less the same thing
on a non-xfs filesystem, so, ah, do that I guess, and re-enable
it.

# quota/xfs_quota -f -x -c path -c free /mnt/test
          Filesystem          Pathname
[000] (F) /mnt/test           /dev/sdb1 (uquota)

Filesystem           1K-blocks       Used  Available  Use% Pathname
/dev/sdb1             20511356      45000   20466356    0% /mnt/test

Fix the short help text for -N while we're at it.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Bill O'Donnell <billodo@redhat.com
Signed-off-by: Dave Chinner <david@fromorbit.com>